+++ e2e/e2etests/test_bitcoin_withdraw_p2sh.go
@@ -11,8 +11,6 @@ func TestBitcoinWithdrawP2SH(r *runner.E2ERunner, args []string) {
 	// check length of arguments
 	require.Len(r, args, 2)
 
-	r.SetBtcAddress(r.Name, false)
-
 	// parse arguments and withdraw BTC
 	defaultReceiver := "2N6AoUj3KPS7wNGZXuCckh8YEWcSYNsGbqd"
 	receiver, amount := parseBitcoinWithdrawArgs(r, args, defaultReceiver)