+++ x/observer/keeper/events.go
@@ -1,9 +1,10 @@
 package keeper
 
 import (
+	"encoding/json"
+	"log"
 	"strconv"
 
-	types2 "github.com/coinbase/rosetta-sdk-go/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	types "github.com/zeta-chain/node/x/observer/types"
@@ -21,11 +22,25 @@ func EmitEventBallotCreated(ctx sdk.Context, ballot types.Ballot, observationHas
 	}
 }
 
+// vendor this code from github.com/coinbase/rosetta-sdk-go/types
+func prettyPrintStruct(val interface{}) string {
+	prettyStruct, err := json.MarshalIndent(
+		val,
+		"",
+		" ",
+	)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	return string(prettyStruct)
+}
+
 func EmitEventKeyGenBlockUpdated(ctx sdk.Context, keygen *types.Keygen) {
 	err := ctx.EventManager().EmitTypedEvents(&types.EventKeygenBlockUpdated{
 		MsgTypeUrl:    sdk.MsgTypeURL(&types.MsgUpdateKeygen{}),
 		KeygenBlock:   strconv.Itoa(int(keygen.BlockNumber)),
-		KeygenPubkeys: types2.PrettyPrintStruct(keygen.GranteePubkeys),
+		KeygenPubkeys: prettyPrintStruct(keygen.GranteePubkeys),
 	})
 	if err != nil {
 		ctx.Logger().Error("Error emitting EventKeygenBlockUpdated :", err)